Updated data type to Decimal

+++ grails-app/domain/uk/co/wonderlane/wlpos/loyalty/MemberTransaction.groovy
@@ -14,8 +14,8 @@ class MemberTransaction {
     int transactionId
     DateTime transactionTimestamp
     String redeemableOffers
-    Double transactionTotal
-    Double transactionDiscount
+    BigDecimal transactionTotal
+    BigDecimal transactionDiscount
     Integer transactionPoints
     Integer transactionStamps
     MemberTransactionStatus status
